fixing dataypes

+++ 2024/day11/main.rs
@@ -11,22 +11,22 @@ where
     result // Return the result of the function
 }
 
-fn count_after_blinking(l_cnt: Vec<i32>, n: i32) -> i32 {
-    let mut l_cnt: HashMap<i32, i32> = l_cnt.into_iter().fold(HashMap::new(), |mut acc, x| {
+fn count_after_blinking(l_cnt: Vec<u64>, n: u64) -> u64 {
+    let mut l_cnt: HashMap<u64, u64> = l_cnt.into_iter().fold(HashMap::new(), |mut acc, x| {
         *acc.entry(x).or_insert(0) += 1;
         acc
     });
 
     for _ in 0..n {
-        let mut new_l: HashMap<i32, i32> = HashMap::new();
+        let mut new_l: HashMap<u64, u64> = HashMap::new();
         for (&s, &count) in &l_cnt {
             if s == 0 {
                 *new_l.entry(1).or_insert(0) += count;
             } else if s.to_string().len() % 2 == 0 {
                 let num_str = s.to_string();
                 let mid = num_str.len() / 2;
-                let n1 = num_str[..mid].parse::<i32>().unwrap();
-                let n2 = num_str[mid..].parse::<i32>().unwrap();
+                let n1 = num_str[..mid].parse::<u64>().unwrap();
+                let n2 = num_str[mid..].parse::<u64>().unwrap();
                 *new_l.entry(n1).or_insert(0) += count;
                 *new_l.entry(n2).or_insert(0) += count;
             } else {
@@ -41,7 +41,7 @@ fn count_after_blinking(l_cnt: Vec<i32>, n: i32) -> i32 {
 
 fn part_1() {
     let contents = include_str!("input.txt");
-    let l: Vec<i32> = contents
+    let l: Vec<u64> = contents
         .split_whitespace()
         .map(|s| s.parse().unwrap())
         .collect();
@@ -51,7 +51,7 @@ fn part_1() {
 
 fn part_2() {
     let contents = include_str!("input.txt");
-    let l: Vec<i32> = contents
+    let l: Vec<u64> = contents
         .split_whitespace()
         .map(|s| s.parse().unwrap())
         .collect();
